update meteor 1.3
