+++ .meteor/.finished-upgraders
@@ -10,3 +10,4 @@ notices-for-facebook-graph-api-2
 1.2.0-meteor-platform-split
 1.2.0-cordova-changes
 1.2.0-breaking-changes
+1.3.0-split-minifiers-package
