+++ .meteor/packages
@@ -12,7 +12,6 @@ session                 # Client-side reactive dictionary for your app
 jquery                  # Helpful client-side library
 tracker                 # Meteor's client-side reactive programming library
 
-standard-minifiers      # JS/CSS minifiers run for production mode
 es5-shim                # ECMAScript 5 compatibility for older browsers.
 ecmascript              # Enable ECMAScript2015+ syntax in app code
 
@@ -26,8 +25,7 @@ kadira:blaze-layout
 fortawesome:fontawesome
 
 
-#<NPM>
-meteorhacks:npm
-cosmos:browserify
-npm-container
-#</NPM>
+
+dbarrett:dropzonejs
+standard-minifier-css
+standard-minifier-js
