+++ .meteor/versions
@@ -1,75 +1,76 @@
+allow-deny@1.0.4
 arboleya:electrify@2.1.4
-autoupdate@1.2.4
-babel-compiler@5.8.24_1
-babel-runtime@0.1.4
-base64@1.0.4
-binary-heap@1.0.4
-blaze@2.1.3
-blaze-html-templates@1.0.1
-blaze-tools@1.0.4
-boilerplate-generator@1.0.4
-caching-compiler@1.0.0
-caching-html-compiler@1.0.2
-callback-hook@1.0.4
-check@1.1.0
-coffeescript@1.0.11
-cosmos:browserify@0.9.3
-ddp@1.2.2
-ddp-client@1.2.1
-ddp-common@1.2.2
-ddp-server@1.2.2
-deps@1.0.9
-diff-sequence@1.0.1
-ecmascript@0.1.6
-ecmascript-runtime@0.2.6
-ejson@1.0.7
-es5-shim@4.1.14
-fastclick@1.0.7
+autoupdate@1.2.9
+babel-compiler@6.6.4
+babel-runtime@0.1.8
+base64@1.0.8
+binary-heap@1.0.8
+blaze@2.1.7
+blaze-html-templates@1.0.4
+blaze-tools@1.0.8
+boilerplate-generator@1.0.8
+caching-compiler@1.0.4
+caching-html-compiler@1.0.6
+callback-hook@1.0.8
+check@1.2.1
+dbarrett:dropzonejs@4.2.1
+ddp@1.2.5
+ddp-client@1.2.7
+ddp-common@1.2.5
+ddp-server@1.2.6
+deps@1.0.12
+diff-sequence@1.0.5
+ecmascript@0.4.3
+ecmascript-runtime@0.2.10
+ejson@1.0.11
+es5-shim@4.5.10
+fastclick@1.0.11
 fortawesome:fontawesome@4.5.0
-geojson-utils@1.0.4
-hot-code-push@1.0.0
-html-tools@1.0.5
-htmljs@1.0.5
-http@1.1.1
-id-map@1.0.4
-jquery@1.11.4
+geojson-utils@1.0.8
+hot-code-push@1.0.4
+html-tools@1.0.9
+htmljs@1.0.9
+http@1.1.5
+id-map@1.0.7
+jquery@1.11.8
 kadira:blaze-layout@2.3.0
-kadira:flow-router@2.10.0
-launch-screen@1.0.4
-livedata@1.0.15
-logging@1.0.8
-meteor@1.1.10
-meteor-base@1.0.1
-meteorhacks:async@1.0.0
-meteorhacks:npm@1.5.0
-minifiers@1.1.7
-minimongo@1.0.10
-mobile-experience@1.0.1
-mobile-status-bar@1.0.6
-mongo@1.1.3
-mongo-id@1.0.1
-npm-container@1.2.0
-npm-mongo@1.4.39_1
-observe-sequence@1.0.7
-ordered-dict@1.0.4
-promise@0.5.1
-random@1.0.5
-reactive-dict@1.1.3
-reactive-var@1.0.6
-reload@1.1.4
-retry@1.0.4
-routepolicy@1.0.6
-session@1.1.1
-spacebars@1.0.7
-spacebars-compiler@1.0.7
-standard-minifiers@1.0.2
-stylus@2.511.1
-templating@1.1.5
-templating-tools@1.0.0
-tracker@1.0.9
+kadira:flow-router@2.12.1
+launch-screen@1.0.11
+livedata@1.0.18
+logging@1.0.12
+meteor@1.1.14
+meteor-base@1.0.4
+minifier-css@1.1.11
+minifier-js@1.1.11
+minimongo@1.0.16
+mobile-experience@1.0.4
+mobile-status-bar@1.0.12
+modules@0.6.1
+modules-runtime@0.6.3
+mongo@1.1.7
+mongo-id@1.0.4
+npm-mongo@1.4.43
+observe-sequence@1.0.11
+ordered-dict@1.0.7
+promise@0.6.7
+random@1.0.9
+reactive-dict@1.1.7
+reactive-var@1.0.9
+reload@1.1.8
+retry@1.0.7
+routepolicy@1.0.10
+session@1.1.5
+spacebars@1.0.11
+spacebars-compiler@1.0.11
+standard-minifier-css@1.0.6
+standard-minifier-js@1.0.6
+stylus@2.511.5
+templating@1.1.9
+templating-tools@1.0.4
+tracker@1.0.13
 twbs:bootstrap@3.3.6
-ui@1.0.8
-underscore@1.0.4
-url@1.0.5
-webapp@1.2.3
-webapp-hashing@1.0.5
+ui@1.0.11
+underscore@1.0.8
+url@1.0.9
+webapp@1.2.8
+webapp-hashing@1.0.9
