+++ client/html/Layouts/MainLayout.html
@@ -1,54 +1,16 @@
 <template name="MainLayout">
   <nav class="navbar navbar-default navbar-fixed-top">
-    <div class="container">
-      <h1>Welcome to MeteorTorrent Alpha</h1>
-    </div>
+      <div class="col-md-4"><a href="/"><h1>VIDEOS</h1></a></div>
+      <div class="col-md-4"><a href="/imagenes"><h1>Imagenes</h1></a></div>
+      <div class="col-md-4"><a href="/add"><h1>+</h1></a></div>
   </nav>
   <div class="container">
     {{> Template.dynamic template=main}}
   </div>
-  <nav class="navbar navbar-default navbar-fixed-bottom">
+  <!-- <nav class="navbar navbar-default navbar-fixed-bottom">
     <div class="container">
       Pausa y otros
     </div>
-  </nav>
-</template>
-
-
-<template name="videoTemp">
-  <div class="col-xs-12">
-  <div id="video-container" >
-		<!-- Video -->
-		<video id="video">
-		  <p>
-		    Your browser doesn't support HTML5 video.
-		    <a href="videos/mikethefrog.mp4">Download</a> the video instead.
-		  </p>
-		</video>
-		<!-- Video Controls -->
-		<div class="container-fluid animated" id="video-controls">
-      <div class="container-fluid">
-				<div class="row">
-          <div class="col-xs-12">
-						<input type="range" class="slider" id="seek-bar" value="0">
-					</div>
-					<div class="col-xs-1">
-						<button type="button" id="play-pause" class="btn btn-xs play"><i class='fa fa-play fa-2x'></i></button>
-					</div>
-					<div class="col-xs-1">
-						<button type="button" class="btn btn-sm" id="mute"><i class='fa fa-volume-up fa-2x'></i></button>
-					</div>
-					<div class="col-xs-3">
-						<input type="range" class="slider" id="volume-bar" min="0" max="1" step="0.01" value="1">
-					</div>
-          <div class="col-xs-4"></div>
-  				<div class="col-xs-3">
-  					<button type="button" id="full-screen"><i class="fa fa-arrows-alt fa-2x"></i></button>
-  			  </div>
-        </div>
-      </div>
-
-		</div>
-		</div>
-   </div>
+  </nav> -->
+  <script src="https://cdn.jsdelivr.net/webtorrent/latest/webtorrent.min.js"></script>
 </template>
