--- client/html/MeteorTorrent.html
@@ -1,44 +1,0 @@
-<!--
-<body>
-  <nav class="navbar navbar-default navbar-fixed-top">
-    <div class="container">
-      <h1>Welcome to MeteorTorrent Alpha</h1>
-    </div>
-  </nav>
-  <div class="container">
-    {{> inputVideo}}
-    <div class="row">
-      {{> torrents}}
-    </div>
-  </div>
-  <nav class="navbar navbar-default navbar-fixed-bottom">
-    <div class="container">
-      Pausa y otros
-    </div>
-  </nav>
-</body>
-
-<template name="torrents">
-  {{#each torrents}}
-    {{> torrent}}
-  {{/each}}
-</template>
-
-<template name="torrent">
-  <div class="video-con col-md-6">
-    <div class="inside">
-      <video id={{this._id}} controls>
-        {{torrente}}
-      </video>
-    </div>
-  </div>
-</template>
-
-<template name="inputVideo">
-  <div class="container-fluid">
-    <form class="new-torrent">
-    <input type="text" name="text" class="col-xs-12" placeholder="Escribe para agregar" />
-    </form>
-  </div>
-</template>
--->
