+++ client/html/Templates.html
@@ -1,0 +1,72 @@
+<template name="inputVideo">
+  <form class="new-torrent" >
+    <div class="form-group">
+      <input type="text" name="name" class="col-xs-12" placeholder="Escribe el nombre" />
+      <input type="text" name="text" class="col-xs-12" placeholder="Escribe el MagnetURL" />
+      <input type="text" name="url" class="col-xs-12" placeholder="Escribe la url de la imagen" />
+    <input type="submit" class="btn btn-primary btn-raised" value="Enviar" >
+    </div>
+  </form>
+</template>
+<template name="torrents">
+  {{#each torrents}}
+    {{> torrent}}
+  {{else}}
+    {{> noTorrents}}
+  {{/each}}
+</template>
+
+<template name="torrent">
+  <a href="/video/{{_id}}">
+  <div class="video-con col-md-6">
+    <div class="inside" style="background-image: url('{{url}}')">
+    </div>
+  </div>
+  </a>
+</template>
+
+<template name="VideoPlayer">
+  <div class="col-xs-12">
+  <div id="video-container" >
+    <!-- Video -->
+    <video id="video">
+      <p>
+        Your browser doesn't support HTML5 video
+      </p>
+    </video>
+    <!-- Video Controls -->
+    <div class="container-fluid animated" id="video-controls">
+      <div class="container-fluid">
+        <div class="row">
+          <div class="col-xs-12">
+            <input type="range" class="slider" id="seek-bar" value="0">
+          </div>
+          <div class="col-xs-1">
+            <button type="button" id="play-pause" class="btn btn-xs play"><i class='fa fa-play fa-2x'></i></button>
+          </div>
+          <div class="col-xs-1">
+            <button type="button" class="btn btn-sm" id="mute"><i class='fa fa-volume-up fa-2x'></i></button>
+          </div>
+          <div class="col-xs-3">
+            <input type="range" class="slider" id="volume-bar" min="0" max="1" step="0.01" value="1">
+          </div>
+          <div class="col-xs-4"></div>
+          <div class="col-xs-3">
+            <button type="button" id="full-screen"><i class="fa fa-arrows-alt fa-2x"></i></button>
+          </div>
+        </div>
+      </div>
+
+    </div>
+    </div>
+   </div>
+</template>
+<template name="add">
+  {{> inputVideo}}
+  {{> dropzone url='http://somewebsite.com/upload' id='dropzoneDiv'}}
+</template>
+
+<template name="noTorrents">
+  No torrents added, please add one
+
+</template>
