+++ client/style.css
@@ -2,6 +2,8 @@
 	width: auto;
 	height: auto;
 	position: relative;
+	background-color: rgba(0, 0, 0, 0.5);
+	padding-bottom: -5px;
 }
 video {
 	width: 100%;
@@ -170,96 +172,3 @@ input::-ms-fill-upper {
 input::-ms-tooltip {
   display: none;
 }
-
-
-/*input[type=range].slider {
-  -webkit-appearance: none;
-  width: 100%;
-  margin: 7.25px 0;
-}
-input[type=range].slider::-moz-range-progress {
-	background: #3371a9;
-	border: 0;
-	margin-top: 0;
-}
-
-input[type=range].slider:focus {
-  outline: none !important;
-}
-input[type=range].slider::-webkit-slider-runnable-track {
-  width: 100%;
-  height: 4.5px;
-  cursor: pointer;
-  box-shadow: 0px 0px 3.9px rgba(0, 0, 0, 0.46), 0px 0px 0px rgba(13, 13, 13, 0.46);
-	background: #666;
-  border-radius: 25px;
-  border: 0.2px solid #010101;
-}
-input[type=range].slider::-webkit-slider-thumb {
-  box-shadow: 0px 0px 1px #000000, 0px 0px 0px #0d0d0d;
-  border: 0.2px solid #000000;
-  height: 19px;
-  width: 20px;
-  border-radius: 50px;
-  background: #ffffff;
-  cursor: pointer;
-  -webkit-appearance: none;
-  margin-top: -7.45px;
-}
-input[type=range].slider:focus::-webkit-slider-runnable-track {
-	background: #666;
-}
-input[type=range].slider::-moz-range-track {
-  width: 100%;
-  height: 4.5px;
-  cursor: pointer;
-  box-shadow: 0px 0px 3.9px rgba(0, 0, 0, 0.46), 0px 0px 0px rgba(13, 13, 13, 0.46);
-  background: #3371a9;
-  border-radius: 25px;
-  border: 0.2px solid #010101;
-}
-input[type=range].slider::-moz-range-thumb {
-  box-shadow: 0px 0px 1px #000000, 0px 0px 0px #0d0d0d;
-  border: 0.2px solid #000000;
-  height: 19px;
-  width: 20px;
-  border-radius: 50px;
-  background: #ffffff;
-  cursor: pointer;
-}
-input[type=range].slider::-ms-track {
-  width: 100%;
-  height: 4.5px;
-  cursor: pointer;
-  background: transparent;
-  border-color: transparent;
-  color: transparent;
-}
-input[type=range].slider::-ms-fill-lower {
-  background: #2e6799;
-  border: 0.2px solid #010101;
-  border-radius: 50px;
-  box-shadow: 0px 0px 3.9px rgba(0, 0, 0, 0.46), 0px 0px 0px rgba(13, 13, 13, 0.46);
-}
-input[type=range].slider::-ms-fill-upper {
-  background: #3371a9;
-  border: 0.2px solid #010101;
-  border-radius: 50px;
-  box-shadow: 0px 0px 3.9px rgba(0, 0, 0, 0.46), 0px 0px 0px rgba(13, 13, 13, 0.46);
-}
-input[type=range].slider::-ms-thumb {
-  box-shadow: 0px 0px 1px #000000, 0px 0px 0px #0d0d0d;
-  border: 0.2px solid #000000;
-  height: 19px;
-  width: 20px;
-  border-radius: 50px;
-  background: #ffffff;
-  cursor: pointer;
-  height: 4.5px;
-}
-input[type=range].slider:focus::-ms-fill-lower {
-  background: #3371a9;
-}
-input[type=range].slider:focus::-ms-fill-upper {
-  background: #387bb9;
-}*/
