+++ client/stylus.styl
@@ -24,8 +24,9 @@ body
   max-height 325px
   width auto
   background-color rgba(255,255,255,.66)
+  background-size cover
   box-shadow 0px 6px 25px 1px rgba(0,0,0,0.75)
-// 
+//
 // video
 //   width 100%
 //   height auto
