--- packages.json
@@ -1,3 +1,0 @@
-{
-  "webtorrent": "0.72.0"
-}
