+++ server/methods.js
@@ -1,10 +1,12 @@
 Meteor.methods({ //News Methods
-  addTorrent :function (magnetURL) {
+  addTorrent :function (magnetURL, name, url) {
     if (magnetURL == "") {
 
     } else {
     Torrents.insert({
       magnetURL: magnetURL,
+      name: name,
+      url: url,
       createdAt: new Date()
     }); }
   },
